Use kubecontainer to handle device
Update bazel file
